requirements.txt
fastapi==0.63.*
uvicorn[standard]==0.13.*
python-jose[cryptography]==3.2.*
python-multipart==0.0.*
